fixed hamming

+++ include/HammingDistance.h
@@ -6,12 +6,12 @@
 
 class HammingDistance : public ImageDistance {
 public:
-    HammingDistance(double thresh);
+    HammingDistance(int thresh);
     double calc(const ImageData &dat1, const ImageData &dat2);
     std::string get_class_name();
 
 protected:
-    double thresh;
+    int thresh;
 };
 
 #endif // HAMMINGDISTANCE_H
